+++ Makefile
@@ -1,0 +1,5 @@
+clean:
+	@echo "Cleaning..."
+	@rm -rf db tmp
+	@echo "## Done ##"
+
