+++ utils/utils.go
@@ -78,3 +78,15 @@ func RemoveFile(filepath string) error {
 
 	return nil
 }
+
+// Ensure that Dir is created
+func EnsureDir(path string) error {
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		fmt.Println("Directory does not exist, creating it...")
+		if err := os.MkdirAll(path, os.ModePerm); err != nil {
+			return fmt.Errorf("failed to create directory: %w", err)
+		}
+	}
+
+	return nil
+}
